+++ Documentation/Release Notes.md
@@ -1,3 +1,9 @@
+# Version 1.0.6
+
+## Documentation
+
+- We added a [tutorial](Tutorials/Release.md) that explains the necessary steps for releasing a new version of ICOc
+
 # Version 1.0.5
 
 ## Configuration